+++ src/sc/src/k8/controllers/spu.rs
@@ -7,7 +7,7 @@ use fluvio_controlplane_metadata::{
 };
 use fluvio_stream_dispatcher::actions::WSAction;
 use fluvio_types::SpuId;
-use k8_client::{ClientError};
+use k8_client::ClientError;
 use tracing::debug;
 use tracing::trace;
 use tracing::error;
@@ -15,7 +15,7 @@ use tracing::instrument;
 
 use fluvio_future::task::spawn;
 use fluvio_future::timer::sleep;
-use k8_types::core::service::LoadBalancerIngress;
+use k8_types::core::service::{LoadBalancerIngress, LoadBalancerType};
 
 use crate::{
     stores::{StoreContext, K8ChangeListener},
@@ -196,34 +196,63 @@ impl SpuController {
         let spu_id = spu_md.key();
         let svc_id = svc_md.key();
 
-        // check if ingress exists
-        let spu_ingress = &spu_md.spec.public_endpoint.ingress;
+        // Check what type of load balancing we're using between NodePort and LoadBalancer
+        //debug!("Applying ingress address based on services");
+        //debug!("SPU: {:?}", &spu_md);
+        //debug!("SVC: {:?}", &svc_md);
 
+        // Get the current ingress on the spu
+        let spu_ingress = spu_md.spec.public_endpoint.ingress.clone();
+        let spu_port = spu_md.spec.public_endpoint.port;
+
+        let spu_ingressport = IngressPort {
+            port: spu_port,
+            ingress: spu_ingress,
+            ..Default::default()
+        };
+
+        // Get the external ingress from the service
+        // Look at svc_md to identify if LoadBalancer
+        let lb_type = svc_md.spec().inner().r#type.as_ref();
+        let spu_svc_port = svc_md.spec().inner().ports[0].port;
+        let spu_svc_nodeport = svc_md.spec().inner().ports[0]
+            .node_port
+            .ok_or_else(|| ClientError::Other("SPU service missing NodePort".into()))?;
+
+        // This will either have a value from External-IP, or will be empty
+        // If empty, the use `fluvio.io/ingress-address` annotation to set an external address for SPU
         let svc_lb_ingresses = svc_md.status.ingress();
-        let mut computed_spu_ingress: Vec<IngressAddr> =
-            svc_lb_ingresses.iter().map(convert).collect();
 
-        if let Some(address) = SpuServiceSpec::ingress_annotation(svc_md.ctx().item()) {
-            if let Ok(ip_addr) = address.parse::<IpAddr>() {
-                computed_spu_ingress.push(IngressAddr {
-                    hostname: None,
-                    ip: Some(ip_addr.to_string()),
-                });
-            } else {
-                computed_spu_ingress.push(IngressAddr {
-                    hostname: Some(address.clone()),
-                    ip: None,
-                });
-            }
-        }
+        // Choose the external port based on the service type
+        let mut computed_spu_ingressport = match lb_type {
+            Some(LoadBalancerType::NodePort) => IngressPort {
+                port: spu_svc_nodeport,
+                ..Default::default()
+            },
+            _ => IngressPort {
+                port: spu_svc_port,
+                ingress: svc_lb_ingresses.iter().map(convert).collect(),
+                ..Default::default()
+            },
+        };
 
-        if &computed_spu_ingress != spu_ingress {
+        // Add additional ingress via annotation value
+        add_ingress_from_svc_annotation(&svc_md, &mut computed_spu_ingressport.ingress);
+
+        debug!(
+            "Computed SPU ingress after applying any svc annotation: {:?}",
+            &computed_spu_ingressport
+        );
+
+        // We're going to come in with a fully built IngressPort
+        if computed_spu_ingressport != spu_ingressport {
             let mut update_spu = spu_md.spec.clone();
             debug!(
                 "updating spu: {} public end point: {:#?} from svc: {}",
-                spu_id, computed_spu_ingress, svc_id
+                spu_id, computed_spu_ingressport, svc_id
             );
-            update_spu.public_endpoint.ingress = computed_spu_ingress;
+            update_spu.public_endpoint.ingress = computed_spu_ingressport.ingress;
+            update_spu.public_endpoint.port = computed_spu_ingressport.port;
             self.spus.create_spec(spu_id.to_owned(), update_spu).await?;
         } else {
             debug!(
@@ -313,6 +342,26 @@ impl SpuController {
         debug!(spu_count, "finished applying spu");
 
         Ok(())
+    }
+}
+
+fn add_ingress_from_svc_annotation(
+    svc_md: &MetadataStoreObject<SpuServiceSpec, K8MetaItem>,
+    computed_spu_ingress: &mut Vec<IngressAddr>,
+) {
+    debug!("Checking ingress for annotations: {:#?}", &svc_md);
+    if let Some(address) = SpuServiceSpec::ingress_annotation(svc_md.ctx().item()) {
+        if let Ok(ip_addr) = address.parse::<IpAddr>() {
+            computed_spu_ingress.push(IngressAddr {
+                hostname: None,
+                ip: Some(ip_addr.to_string()),
+            });
+        } else {
+            computed_spu_ingress.push(IngressAddr {
+                hostname: Some(address.clone()),
+                ip: None,
+            });
+        }
     }
 }